+++ packages/incipit-base/changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Add safe versions of `Fractional` division methods.
+
 # 0.2.0.0
 
 * Add safe, `Maybe` returning, versions of the methods of `Integral`.